Remove redundant -t when extracting thumbnails.
The FFmpeg documentation makes it sound like -vframes and -t are
mutually exclusive:
"""
If you want to extract just a limited number of frames, you can use the
above command in combination with the -vframes or -t option, or in
combination with -ss to start extracting from a certain point in time.
"""
http://ffmpeg.org/ffmpeg.html

...and I had one video where no output file was produced (and no error
reported) when both -vframes and -t were specified.