Use invariant culture for formatting Double values in command line
On systems where the decimal separator is a , instead of . MediaToolkit
will not work correctly because it does not use invariant culture to
format Double values, generating command line arguments like "-ss 923,52"
which ffmpeg will reject with the error "Invalid duration specification
for ss: 923,52". Using the invariant culture to format the arguments fixes
this.

Signed-off-by: Axel Gembe <axel@gembe.net>